Night summaries to dataframes through indexer
Night summaries were previously turned into dataframes through splitting 
lines on '|' and then removing spaces. This is a problem when:

1) A row is missing a value, such as the fileroot
2) A value has a space in it, such as a PI or comment

The new method for parsing a night summary is to identify the 
locations of the '|' in the first row, which becomes the header of 
the dataframe. Then each subsequent line of the summary is 
split at the location of those indices. This keeps the number of 
values in each row the same as the number of columns in the 
header. The parsing of the summary is done in the new function 
convert_summary_to_dataframe. Both the download_night_summary 
and download_allnight_summary functions call the 
convert_summary_to_dataframe function.

A None is returned when either the night summary 
does not exist or is blank. The NightSummary class 
follows up on this self.data_all = None by setting 
self.data = None as well.